requirements.txt
Django==1.9.11
Werkzeug==0.9.6

# django-cors-headers is completely optional and probably makes sense only in development
django-cors-headers==0.12

django-extensions==1.7.4
django-pipeline==1.6.9
djangorestframework==3.5.2

drf-compound-fields==0.2.1
ipython==2.1.0
python-dateutil==2.2
pytz==2014.4
